Pure Xcode: Supplement vfs with -F (#12)
Pure Xcode: Supplement vfs with -F

Logically `-F$(BUILT_PRODUCTS_DIR)` include provides the identical
functionality as the VFS does in Bazel mode. In other words, the include
and VFS yield the same behavior in clang and swift. WIP state is because
we need make this Pure Xcode only ( actually want it in Bazel mode )

Finally, this allows us to:
- remove https://github.com/bazel-ios/rules_ios/pull/470 altogether
- always compile with `apple.virtualize_frameworks` for CLI builds and
  Xcode builds in Pure Xcode mode